+++ noodles-bam/src/record/data/fields.rs
@@ -1,25 +1,23 @@
 use std::io::{self, BufRead};
 
-use crate::record::data::Field;
-
-use super::Reader;
+use crate::{reader::record::data::read_field, record::data::Field};
 
 /// An iterator over BAM record data fields.
 ///
-/// This is created by calling [`Reader::fields`].
+/// This is created by calling [`crate::record::Data::fields`].
 pub struct Fields<R>
 where
     R: BufRead,
 {
-    reader: Reader<R>,
+    inner: R,
 }
 
 impl<R> Fields<R>
 where
     R: BufRead,
 {
-    pub(crate) fn new(reader: Reader<R>) -> Self {
-        Self { reader }
+    pub(crate) fn new(inner: R) -> Self {
+        Self { inner }
     }
 }
 
@@ -30,7 +28,7 @@ where
     type Item = io::Result<Field>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        match self.reader.read_field() {
+        match read_field(&mut self.inner) {
             Ok(Some(field)) => Some(Ok(field)),
             Ok(None) => None,
             Err(e) => Some(Err(e)),